+++ src/routes/settings/+page.svelte
@@ -24,6 +24,7 @@
 	const dayThemes = [
 		{ themeClass: 'vanilla', displayName: 'Plain Vanilla' },
 		{ themeClass: 'strawberry', displayName: 'Strawberry Sundae' },
+		{ themeClass: 'lemon', displayName: 'Lemon Tart' },
 	];
 
 	const nightThemes = [